Added a page for Inventory table
