+++ src/Pages/InventoryListPage/InventoryListPage.js
@@ -1,0 +1,41 @@
+import React from "react";
+import "./InventoryListPage.scss";
+import PageWrapper from "../../Components/PageWrapper/PageWrapper";
+import InventoryTable from "../../Components/InventoryTable/InventoryTable";
+import axios from 'axios';
+import { useEffect, useState } from "react";
+
+const InventoryListPage = () => {
+    const [inventory, setInventory] = useState([]);
+    const [loading, setLoading] = useState(false);
+  
+    useEffect(() => {
+      const fetchInventory = async () => {
+        setLoading(true);
+        try {
+          const { status, data } = await axios.get(
+            `${process.env.REACT_APP_BACKEND_URL}/inventories`
+          );
+          if (status === 200) {
+            setInventory(data);
+          } else {
+            console.error('Failed to fetch inventory:', status);
+          }
+        } catch (error) {
+          console.error('Error fetching inventory:', error);
+        } finally {
+          setLoading(false);
+        }
+      };
+  
+      fetchInventory();
+    }, []);
+  
+    return (
+      <PageWrapper title={"Inventory"} location={"Inventory"}>
+        <InventoryTable inventory={inventory} loading={loading} /> {/* Pass inventory and loading as props */}
+      </PageWrapper>
+    );
+  };
+  
+  export default InventoryListPage;
